+++ semestralka/src/Player.h
@@ -45,7 +45,7 @@ public:
         deck.shuffle();
     }
     void print(unsigned int flag) const override; //if flag is zero, censor the hand, if it is 1, print it
-    bool playCard(int NO, Player& enemyPlayer); //returns false if the board is full.
+    bool playCard(unsigned long NO, Player& enemyPlayer); //returns false if the board is full.
     bool attack(int attacking, int defending, Player& second); //a complex function for a player attacking another player with one of his cards
     void doBattlecry(int batID, Player& enemyPlayer); //A battlecry card has been played, and this function will carry out the battlecry effect on the game
     friend ostream& operator <<(ostream& out, const Player& p){
